made index the LoginScreen
