+++ app/(home)/ReportScreen.js
@@ -10,7 +10,7 @@ import * as ImagePicker from 'expo-image-picker';
 import { database } from '../../backend/config/firebaseConfig';
 import BackButtonComponent from "../../assets/svg/back_button.js";
 import React, { useState, useEffect } from "react";
-import { useLocalSearchParams, router } from 'expo-router';
+import { useRouter } from 'expo-router';
 
 //import { useRoute } from '@react-navigation/native';
 
@@ -20,10 +20,10 @@ import emailjs from 'emailjs-com';
 
 export default function ReportScreen() {
 
-    const params = useLocalSearchParams();
-    let listing = params.listing;
+    const route = useRouter();
+    const listing = route.listing;
 
-    console.log(listing.images);
+    console.log(listing);
 
     // Sending the email to the admin.
     const [emailSent, setEmailSent] = useState(false);
